main thing is new custom hook for filtering data on homepage by date, but also changes to quick selectbox
