+++ src/components/Home/Map/IncidentFocus.js
@@ -3,23 +3,13 @@ import { CaretRightOutlined } from '@ant-design/icons';
 import { useSelector } from 'react-redux';
 import { Collapse, List } from 'antd';
 import IncidentFocusCard from './IncidentFocusCard';
+import useIncidentFilter from '../../../hooks/useIncidentFilter';
+
 const { Panel } = Collapse;
 
 //useSelectors for getting state for map
 export default function IncidentFocus({ zoomOnCluster }) {
-  const activeFocus = useSelector(state => state.map.focus.active);
-  const clusterList = useSelector(state => state.map.focus.cluster.list);
-  const queryList = useSelector(state => state.map.focus.query.list);
-  const defaultList = useSelector(state => [state.incident.timeline[0]]); // most recent incident
-
-  //Focuses the map into specified region
-  const activeList = useMemo(() => {
-    return activeFocus === 'cluster'
-      ? clusterList
-      : activeFocus === 'query'
-      ? queryList
-      : defaultList;
-  }, [activeFocus, clusterList, queryList, defaultList]);
+  const filteredIncidents = useIncidentFilter();
 
   return (
     <div className="map-menu-background">
@@ -45,7 +35,7 @@ export default function IncidentFocus({ zoomOnCluster }) {
           >
             <div className="incident-content">
               <List>
-                {activeList.map(id => (
+                {filteredIncidents.map(id => (
                   <IncidentFocusCard
                     id={id}
                     key={id}
